fix compile errors and contextLoads failing
- reorder annotations in entity. New Order block of lombok annotations -> Block of jpa.persistence annotations above class
- add @EqualsAndHashCode(callSuper = true) for all entities
fixes compile errors and possible feature issues because super is not called
- remove some comments in entities
- remove direct assignments to private variables
fixes compile error, probably not needed as well and @Builder would ignore them that way as well
- remove @Import(TestcontainersConfiguration.class) in all integration tests, since redundant
- make @SpringBootTest to use random Port in Sj2324SeedprojectApplicationTests
- move postgresContainer bean from TestMainApplication to TestContainerConfiguration
- make main in TestMainApplication use TestContainerConfiguration
- remove on spring-boot-starter-thymeleaf as dependency, since it was listed twice in pom.xml
- bump spring-boot-starter-parent to version 3.2.3 from 3.2.4
solving a compile error related to repackage not supporting class file major version 66
- make postgresql from testcontainers be in runtime scope
fixes not being able to verify package with maven, since driver could not be found

fixes #12
