+++ src/main/java/at/spengergasse/sj2324seedproject/domain/Address.java
@@ -15,9 +15,9 @@ import org.springframework.data.jpa.domain.AbstractPersistable;
 
 // implemented by MM
 @Data
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
 
 
 @Embeddable
