+++ src/main/java/at/spengergasse/sj2324seedproject/domain/Producer.java
@@ -6,38 +6,33 @@ import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.domain.AbstractPersistable;
 
 import java.util.List;
 
 @Data
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-@Table(name = "producers")
-@Entity
-public class Producer extends AbstractPersistable<Long>{
+@EqualsAndHashCode(callSuper = true)
 
-    /*
-    Relations
-     */
+
+@Entity
+@Table(name = "producers")
+public class Producer extends AbstractPersistable<Long>{
 
     @OneToMany(fetch = FetchType.LAZY, cascade = {CascadeType.PERSIST, CascadeType.MERGE})
     private List<StorageObjectMeta> storageObjectMeta;
 
-
-    /*
-     Attributes
-      */
-
     @Column(name = "shortname", length = ConstantsDomain.DEFAULT_LENGTH)
     @NotBlank
-    private String shortname = ConstantsDomain.DEFAULT_VALUE;
+    private String shortname;
 
     @NotBlank
     @Column(name = "producer_name", length =  ConstantsDomain.DEFAULT_LENGTH)
-    private String name = ConstantsDomain.DEFAULT_VALUE;
+    private String name;
