+++ src/main/java/at/spengergasse/sj2324seedproject/domain/Profile.java
@@ -11,9 +11,11 @@ import org.hibernate.validator.constraints.Length;
 
 @Data
 @Builder
-@Embeddable
 @NoArgsConstructor
 @AllArgsConstructor
+
+
+@Embeddable
 public class Profile {
 
     private static final int USERNAME_LENGTH = 10;
