+++ src/main/java/at/spengergasse/sj2324seedproject/domain/Reservation.java
@@ -15,14 +15,18 @@ import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.domain.AbstractPersistable;
 
 @Data
-@Entity
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
+
+
+@Entity
 @Table(name = "reservations")
 public class Reservation extends AbstractPersistable<Long> {
