+++ src/main/java/at/spengergasse/sj2324seedproject/domain/StorageObjectMeta.java
@@ -8,6 +8,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.domain.AbstractPersistable;
 
@@ -16,17 +17,14 @@ import java.util.List;
 
 
 @Data
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
+@EqualsAndHashCode(callSuper = true)
+
 @Entity
 @Table(name = "storage_object_metas")
-
 public class StorageObjectMeta extends AbstractPersistable<Long>{
-
-    /*
-    Relations
-     */
 
     @OneToMany(fetch = FetchType.LAZY, cascade = {CascadeType.MERGE, CascadeType.PERSIST})
     public List<StorageObject> storageobject;
@@ -40,17 +38,13 @@ public class StorageObjectMeta extends AbstractPersistable<Long>{
     @Enumerated(EnumType.STRING)
     private Type type;
 
-    /*
-    Attributes
-     */
-
     @NotBlank
     @Column(name = "storage_object_name")
-    private String name = ConstantsDomain.DEFAULT_VALUE;
+    private String name;
 
     @NotBlank
     @Column(name = "os_version")
-    private String osVersion = ConstantsDomain.DEFAULT_VALUE;
+    private String osVersion;
 
     @Min(0)
     @Column(name = "consumables_per_box")
@@ -62,11 +56,11 @@ public class StorageObjectMeta extends AbstractPersistable<Long>{
 
     @NotBlank
     @Column(name = "wave_length")
-    private String wavelength = ConstantsDomain.DEFAULT_VALUE;
+    private String wavelength;
 
     @NotBlank
     @Column(name = "interface_speed")
-    private String interfacespeed = ConstantsDomain.DEFAULT_VALUE;
+    private String interfacespeed;
 
 
 }
