+++ src/main/java/at/spengergasse/sj2324seedproject/domain/Type.java
@@ -4,8 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
-//@NoArgsConstructor
-@AllArgsConstructor //Error wenn erzeugt  -> java: Konstruktor Type() ist bereits in Enumeration at.spengergasse.sj23247abcif.domain.Type definiert
+@AllArgsConstructor
 public enum Type{
 
     IP_PHONE("IP Phone", "IPP"),
@@ -14,17 +13,6 @@ public enum Type{
     SFP("Sfp", "Sfp"),
     CONSUMABLES("Consumables", "Cs");
 
-
-
-    private String longVersion;
-    private String shortVersion;
-
-//public enum Type {
-//
-//    IP_PHONE,
-//    ROUTER,
-//    SWITCH,
-//    SFP,
-//    CONSUMABLES
-//
+    final private String longVersion;
+    final private String shortVersion;
 }
