+++ src/test/java/at/spengergasse/sj2324seedproject/Sj2324SeedprojectApplicationTests.java
@@ -1,10 +1,14 @@
 package at.spengergasse.sj2324seedproject;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
-//@SpringBootTest
-//@Import(TestContainerConfiguration.class)
-public class Sj2324SeedprojectApplicationTests{
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class Sj2324SeedprojectApplicationTests {
 
     @Test
     void contextLoads(){
