+++ src/test/java/at/spengergasse/sj2324seedproject/persistence/RepositoryStorageObjectTest.java
@@ -13,7 +13,6 @@ import java.util.Objects;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
-@Import(TestcontainersConfiguration.class)
 public class RepositoryStorageObjectTest{
 
     @Autowired
