+++ src/test/java/at/spengergasse/sj2324seedproject/persistence/ReservationRepositoryTest.java
@@ -20,7 +20,6 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 
 @DataJpaTest
-@Import(TestcontainersConfiguration.class)
 class ReservationRepositoryTest {
 
   @Autowired
